# apps/backend/vision-service/docs/openapi.yml
openapi: 3.1.0
info:
  title: Vision Service API
  description: |
    Training runs, their epochs and results, and what is built on them: test results, benchmarks,
    comparisons, visualizations and written findings. New here? Start with the
    [quickstart](/docs/quickstart).

    **Credentials.** Training scripts use a *pipeline key*: a personal API key limited to one project.
    Other personal API keys can reach projects their owner may access, narrowed by their scopes.
    Send either in `Authorization: Bearer <token>`.

    **Responses.** Every path starts with `/api`. A request that works answers
    `{ "success": true, "data": … }`; one that fails answers `{ "success": false, "error": …, "message": … }`
    with the matching status.

    **Privacy.** Projects default to private. A private project, its configurations and results are visible
    to its owning person or owning group’s current members, and to members of its editor groups.
    Public projects allow anonymous reads. Current group roles determine who can contribute, manage or transfer.
  version: 1.0.0
  contact:
    name: Visin Team
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:4010/api
    description: Development server
  - url: https://{host}/api
    description: Your deployment's vision-service
    variables:
      host:
        default: vision-api.example.com
        description: Public hostname of vision-service (VISION_API_URL)
...
components:
  schemas:
    Error:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
          description: The kind of failure
          example: NotFoundError
        message:
          type: string
          description: What went wrong, naming the field or record
          example: 'Training not found with uuid: 5f0c6a2e-1d4b-4c3a-9a53-2b7e1f0d8c11'
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
      required:
        - success
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 3
    Training:
      type: object
      properties:
        _id:
          type: string
        uuid:
          type: string
        name:
          type: string
        description:
          type: string
        datasetId:
          type: string
        configId:
          type: string
          description: Configuration belonging to this training’s project.
        projectId:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        tags:
          type: array
          items:
            type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        metrics:
          type: object
          properties:
            totalTime:
              type: number
            epochCount:
              type: integer
            maxEpoch:
              type: integer
            lastEpochTimestamp:
              type: string
            cpuCost:
              type: number
            gpuCost:
              type: number
            totalCost:
              type: number
    Epoch:
      type: object
      properties:
        _id:
          type: string
        trainingId:
          type: string
        training_uuid:
          type: string
        epoch_uuid:
          type: string
        epoch:
          type: integer
        timestamp:
          type: string
          format: date-time
        results:
          type: object
          description: What the epoch reported, as sent. Conditions, classes and metrics are discovered from it.
        learning_rate:
          type: number
        epoch_time:
          type: number
          description: Seconds the epoch took
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    BearerCredential:
      type: http
      scheme: bearer
      description: |
        A user API key (`vsn_live_…`) or an assistant's OAuth access token, sent as `Authorization: Bearer <key>`.
        A key can be limited to one project, which is what a training pipeline uses: it then reaches only that
        project, runs it creates land there whatever project the request names, and it cannot manage projects.
    SessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: The signed-in browser session. Not for scripts.


  # Referenced throughout `paths` — every error goes through the shared
  # errorHandler, so they all share the Error shape.
  responses:
    BadRequest:
      description: The request was rejected by validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No credential, or one that is invalid, expired or revoked
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The credential is valid but may not do this
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error:
      description: The request failed; the status and message say why
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /trainings:
    get:
      tags: [Trainings]
      summary: Get all trainings
      operationId: getTrainings
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
        - {}
      parameters:
        - name: page
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/page }
        - name: limit
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/limit }
        - name: search
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/search }
        - name: status
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/status }
        - name: datasetId
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/datasetId }
        - name: projectId
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/projectId }
        - name: access
          in: query
          description: '`contribute` keeps only runs in projects you can write to, leaving out other people''s public work.'
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/access }
        - name: tags
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/tags }
        - name: excludeTags
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/excludeTags }
        - name: ids
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/ids }
        - name: sortBy
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/sortBy }
        - name: order
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingsQuery/properties/order }
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      trainings:
                        type: array
                        items:
                          $ref: '#/components/schemas/Training'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        default:
          $ref: '#/components/responses/Error'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Trainings]
      summary: Create a new training
      operationId: createTraining
      description: |
        Start a run before posting its epochs. Send your own `uuid` to refer to the
        run by an id your script already knows; one is generated otherwise.

        Every run belongs to a project: `projectId` (its id or slug) is required, except
        with a credential limited to a project, where the run goes into that project and
        `projectId` is ignored. A project that does not exist, or that you cannot see,
        answers 404 with its name; create it under Projects first. Adding a run takes
        `contribute` on the project.
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./generated/request-schemas.json#/CreateTrainingBody
      responses:
        '201':
          description: Training created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Training'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
  /trainings/stats:
    get:
      tags: [Trainings]
      summary: Get training statistics
      operationId: getTrainingStats
      description: |
        Cost estimates use each project's current CPU/GPU rates. Amounts are
        grouped by currency without exchange-rate conversion. Legacy scalar cost
        fields and currency are present only when exactly one currency is priced.
        Coverage includes unpriced trainings and measured durations in seconds;
        zero rates count as priced. These are current estimates, not historical bills.
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
        - {}
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      totalTrainings:
                        type: integer
                      totalTime:
                        type: number
                      totalEpochs:
                        type: integer
                      avgEpochTime:
                        type: number
                      totalCpuCost:
                        type: number
                      totalGpuCost:
                        type: number
                      totalCost:
                        type: number
                      currency:
                        type: string
                        description: ISO currency code of the optional scalar costs
                      costTotalsByCurrency:
                        type: array
                        description: One subtotal per currency, sorted by currency code; empty if unpriced
                        items:
                          type: object
                          required: [currency, totalCpuCost, totalGpuCost, totalCost]
                          properties:
                            currency:
                              type: string
                              example: EUR
                            totalCpuCost:
                              type: number
                            totalGpuCost:
                              type: number
                            totalCost:
                              type: number
                      costCoverage:
                        type: object
                        required: [pricedTrainings, unpricedTrainings, pricedTime, unpricedTime]
                        properties:
                          pricedTrainings:
                            type: integer
                          unpricedTrainings:
                            type: integer
                          pricedTime:
                            type: number
                            description: Measured seconds with configured rates
                          unpricedTime:
                            type: number
                            description: Measured seconds without configured rates
                      filters:
                        type: object
        default:
          $ref: '#/components/responses/Error'
        '400':
          $ref: '#/components/responses/BadRequest'
      parameters:
        - name: status
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingStatsQuery/properties/status }
        - name: datasetId
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingStatsQuery/properties/datasetId }
        - name: projectId
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingStatsQuery/properties/projectId }
        - name: tags
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingStatsQuery/properties/tags }
  /trainings/tags:
    get:
      tags: [Trainings]
      summary: List the tags in use
      operationId: getTrainingTags
      description: Every tag on a training the caller can see, for filter suggestions.
      responses:
        '200':
          description: Tag names
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { type: string }
        default:
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
        - {}
  /trainings/compare:
    post:
      tags: [Trainings]
      summary: Compare trainings
      operationId: compareTrainings
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./generated/request-schemas.json#/CompareTrainingsBody
      responses:
        '200':
          description: Comparison results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
        default:
          $ref: '#/components/responses/Error'
        '400':
          $ref: '#/components/responses/BadRequest'
  /trainings/{id}:
    get:
      tags: [Trainings]
      summary: Get training by ID
      operationId: getTrainingById
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Training'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [Trainings]
      summary: Update training
      operationId: updateTraining
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./generated/request-schemas.json#/UpdateTrainingBody
      responses:
        '200':
          description: Training updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Training'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Trainings]
      summary: Delete training
      operationId: deleteTraining
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Training deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        default:
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /trainings/deleted:
    get:
      tags: [Trainings]
      summary: List deleted trainings the caller can restore
      operationId: getDeletedTrainings
      description: >
        Most recently deleted first: the caller’s runs in projects they contribute to, and all runs in projects they manage. Requires a signed-in caller.
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
      parameters:
        - name: page
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetDeletedTrainingsQuery/properties/page }
        - name: limit
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetDeletedTrainingsQuery/properties/limit }
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      trainings:
                        type: array
                        items:
                          $ref: '#/components/schemas/Training'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
                          pages:
                            type: integer
        default:
          $ref: '#/components/responses/Error'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /trainings/{id}/restore:
    post:
      tags: [Trainings]
      summary: Restore a deleted training
      operationId: restoreTraining
      description: >
        Brings the training back with the epochs and test results its delete removed. Rows deleted separately beforehand stay deleted, and the training is not re-added to comparisons it was removed from.
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Training restored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Training'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /trainings/uuid/{uuid}:
    get:
      tags: [Trainings]
      summary: Get training by UUID
      operationId: getTrainingByUuid
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
        - {}
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Training'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /trainings/{id}/epochs:
    get:
      tags: [Trainings]
      summary: Get epochs for a training
      operationId: getTrainingWithEpochs
      security:
        - { BearerCredential: [] }
        - { SessionCookie: [] }
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: sortBy
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingWithEpochsQuery/properties/sortBy }
        - name: order
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingWithEpochsQuery/properties/order }
        - name: sample
          in: query
          schema: { $ref: ./generated/request-schemas.json#/GetTrainingWithEpochsQuery/properties/sample }
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      epochs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Epoch'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        default:
          $ref: '#/components/responses/Error'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...